Migrated maxDistanceLimit to int in order to make the filter work

+++ app/src/main/java/com/bham/restaurantapp/Globals.java
@@ -2,7 +2,7 @@ package com.bham.restaurantapp;
 
 public abstract class Globals {
     public static final int DEFAULT_MIN_RATING = 3;
-    public static final float DEFAULT_MAX_DISTANCE_LIMIT = 3;
+    public static final int DEFAULT_MAX_DISTANCE_LIMIT = 3;
     public static final int DEFAULT_BUSINESS_TYPE_POSITION = 0;
     public static final int DEFAULT_REGION_POSITION = 0;
     public static final int DEFAULT_AUTHORITY_POSITION = 0;
